requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0

